requirements.txt
opencv-python==4.5.5
matplotlib==3.5.1
Pillow==9.1.0